pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "excel-intelligent-agent"
version = "0.1.0"
description = "Excel Intelligent Agent System based on Google ADK framework"
authors = [{name = "Excel Agent Team"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "google-adk>=0.1.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "xlrd>=2.0.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "httpx>=0.24.0",
    "requests>=2.31.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "typing-extensions>=4.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
strict = true